Refactor tabled_functions into two files - tabled_functions and tabled_call_executor
