+++ executor/read/mod.rs
@@ -28,8 +28,9 @@ pub mod expression_executor;
 mod immediate_executor;
 mod nested_pattern_executor;
 pub(crate) mod pattern_executor;
-mod step_executor;
-pub(crate) mod tabled_functions;
+pub(crate) mod step_executor;
+pub(crate) mod tabled_call_executor;
+pub mod tabled_functions;
 
 // And use the below one instead
 pub(super) fn TODO_REMOVE_create_executors_for_match(
